fix(@angular/pwa): move manifest and commands where they belong
